% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/days_to_ny.R
\name{days_to_ny}
\alias{days_to_ny}
\title{Days to New Years}
\usage{
days_to_ny(inputdate)
}
\arguments{
\item{inputdate}{Date}
}
\value{
Outputs "Days to New Years for the input date(s):" + a vector of days to New Years
}
\description{
Days to New Years
}
\examples{
days_to_ny(ymd("2020-10-22"))
days_to_ny(as_date(lubridate::ymd("2020-10-22"):ymd("2020-10-29")))

}
